feat: Basic functions completed
