+++ daily/src/constants/storage.ts
@@ -16,6 +16,8 @@ export const SessionStorage = Object.freeze({
 
   /** 当前选择的文档ID (string) */
   VIEW_DOCS_ID: 'view.docsId',
+  /**  当前选择的itemID（string） */
+  VIEW_ITEM_ID: 'view.itemId'
 } as const);
 
 /**
