+++ daily/src/page/home/Home.vue
@@ -1,6 +1,6 @@
 <script setup>
 import ButtonBar from '@/page/home/components/button/ButtonBar/ButtonBar.vue'
-import Banner from '@/page/home/components/banner.vue'
+import Banner from '@/page/home/typeDaily/banner.vue'
 
 import { useI18n } from 'vue-i18n'
